chore: rename e-note-ion.py to scheduler.py
Enables standard import (removing importlib workaround in tests), adds
[project.scripts] for pip-installable CLI, and unblocks pyright module
resolution. Resolves #76.
